Split out EarthModelInterpretation class
