+++ resqpy/organize/OrganizationFeature.py
@@ -1,15 +1,3 @@
-"""organize_old.py: RESQML Feature and Interpretation classes."""
-
-version = '9th November 2021'
-
-# most feature and interpretation classes catered for here
-# stratigraphic classes in strata.py
-
-import logging
-
-log = logging.getLogger(__name__)
-log.debug('organize_old.py version ' + version)
-
 from .organize_functions import alias_for_attribute, equivalent_extra_metadata
 import resqpy.olio.uuid as bu
 import resqpy.olio.xml_et as rqet
