+++ resqpy/organize/__init__.py
@@ -4,3 +4,17 @@ __all__ = [
 
 from .OrganizationFeature import OrganizationFeature
 from .GeobodyFeature import GeobodyFeature
+from .BoundaryFeature import BoundaryFeature
+
+# from .FrontierFeature import FrontierFeature
+# from .GeologicUnitFeature import GeologicUnitFeature
+# from .FluidBoundaryFeature import FluidBoundaryFeature
+# from .RockFluidUnitFeature import RockFluidUnitFeature
+# from .TectonicBoundaryFeature import TectonicBoundaryFeature
+# from .GeneticBoundaryFeature import GeneticBoundaryFeature
+# from .WellboreFeature import WellboreFeature
+# from .FaultInterpretation import FaultInterpretation
+from .EarthModelInterpretation import EarthModelInterpretation
+# from .HorizonInterpretation import HorizonInterpretation
+# from .GeobodyBoundaryInterpretation import GeobodyBoundaryInterpretation
+# from .GeobodyInterpretationllboreInterpretation import GeobodyInterpretationllboreInterpretation
