+++ tests/test_strata.py
@@ -10,6 +10,7 @@ import resqpy.model as rq
 import resqpy.olio.uuid as bu
 import resqpy.organize.organize_old as rqo
 import resqpy.strata as strata
+import resqpy.organize as rqo_new
 
 # the following creates a fully related set of objects, saves to persistent storage, and retrieves
 
@@ -25,7 +26,7 @@ def test_strata(tmp_path):
     emf = rqo.OrganizationFeature(model, feature_name = 'strata model', organization_kind = 'earth model')
     emf.create_xml()
 
-    emi = rqo.EarthModelInterpretation(model, organization_feature = emf)
+    emi = rqo_new.EarthModelInterpretation(model, organization_feature = emf)
     emi.create_xml()
 
     # stratigraphic organisational objects
